[refactor] 스웨거 설정파일 일원화
