+++ src/main/java/potatoes/server/config/SwaggerConfig.java
@@ -3,6 +3,7 @@ package potatoes.server.config;
 import java.util.ArrayList;
 import java.util.List;
 
+import org.springdoc.core.models.GroupedOpenApi;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.MediaType;
@@ -14,6 +15,8 @@ import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.info.Info;
 import io.swagger.v3.oas.models.security.SecurityRequirement;
 import io.swagger.v3.oas.models.security.SecurityScheme;
+import potatoes.server.controller.MailController;
+import potatoes.server.controller.UserController;
 
 @Configuration
 public class SwaggerConfig implements WebMvcConfigurer {
@@ -43,5 +46,25 @@ public class SwaggerConfig implements WebMvcConfigurer {
 		Info info = new Info().title("talking-potatoes-api").description("").version("1.0");
 
 		return new OpenAPI().components(components).info(info).addSecurityItem(securityRequirement);
+	}
+
+	@Bean
+	public GroupedOpenApi getMailApi() {
+		return GroupedOpenApi.builder()
+			.group("mail")
+			.addOpenApiMethodFilter(method ->
+				method.getDeclaringClass().equals(MailController.class))
+			.packagesToScan("potatoes.server.controller")
+			.build();
+	}
+
+	@Bean
+	public GroupedOpenApi getAuthApi() {
+		return GroupedOpenApi.builder()
+			.group("auth")
+			.addOpenApiMethodFilter(method ->
+				method.getDeclaringClass().equals(UserController.class))
+			.packagesToScan("potatoes.server.controller")
+			.build();
 	}
 }
