--- src/test/java/potatoes/server/controller/MailControllerTest.java
@@ -1,4 +1,0 @@
-package potatoes.server.controller;
-
-public class MailControllerTest {
-}
